Allow to pass block to select helper in Rails >= 4.1.0
Beginning with Rails 4.1.0, the select method allows to pass a block to
render as the options string:

http://apidock.com/rails/v4.1.8/ActionView/Helpers/FormOptionsHelper/select
(the link points to 4.1.8, but it is the same for 4.1.0)

Adapt the gem to allow this.

The test is marked to run on Rails >= 4.1.0 only.

One problem remains though: When updating Rails to 4.1.0 in test
environment, many other tests fail because the order of attributes in
the output changed. I don't know how to solve this yet without rewriting
the expected output for different Rails versions (wich is a cumbersome
task ...).