Prevent mutating (possibly) frozen strings
Currently the `#form_group` method attempts to mutate the
`:control_col` option if no label is supplied. Since this option is
passed by the user, it may be a frozen string literal. Instead, we
simply avoid mutating the supplied string.

As we approach Ruby 3 and frozen string literals become more common,
this is more likely to fail.