Fixing input validation messaging.
In bootstrap 4 the classes for the input validations have changed. The
'form-group' div is now expected to have the 'has-danger' class. The
message text is now supposed to have an 'form-control-feedback' class
as opposed to an 'invalid-feedback' class.

This also includes an update to the README to reflect this change.