Issue #444 - Help text rendered in wrong place.
The help text and error message are no longer rendered in the same
place. The help text is now rendered in the `form-group` div instead
of in the `input-group` div. Before we could render the error and help
text together, however that has changed and they need to be separate.

With this separation we need to check in the `generate_help` method to
see if there are errors and if inline errors are being rendered; as we
don't want to render the help text then.