Fixes #443 - Updates `static_control` method.
In Bootstrap 4 the `form-control-static` class no longer exists. It
has been replaced with using the `form-control-plaintext` class on a
readonly input. This reflects that change in approach.

There are a couple of potential problems/oddities.

From a backwards-compatabilty standpoint, if people have been putting
other tags inside of the block this will likely break whatever they
may have been trying to do. There doesn't seem to be any way around
this limitation as it is caused by the text input approach.

Secondly, this approach seems to break convention in this project a
little. It doesn't seem as if any of the `_with_bootstrap` methods are
directly referenced anywhere. I'm not sure how much of a problem this
is, it seemed like the best way to stay DRY.