Removes passing a block as an option from `static_control`.
Since the block will no longer behave the same as it did in previous
versions, it is probably best to remove it entirely. This way tests
will fail hard and make it easier to catch instead of silently behaving
differently.

This also will not overwrite a `control_class` if it is passed in as
an option by the user. There is a new test added to ensure this.

Updates to the CHANGELOG to reflect the breaking changes to this method.