Fixes #443 - Updates `static_control` method. (#446)
* Fixes #443 - Updates `static_control` method.

In Bootstrap 4 the `form-control-static` class no longer exists. It
has been replaced with using the `form-control-plaintext` class on a
readonly input. This reflects that change in approach.

There are a couple of potential problems/oddities.

From a backwards-compatabilty standpoint, if people have been putting
other tags inside of the block this will likely break whatever they
may have been trying to do. There doesn't seem to be any way around
this limitation as it is caused by the text input approach.

Secondly, this approach seems to break convention in this project a
little. It doesn't seem as if any of the `_with_bootstrap` methods are
directly referenced anywhere. I'm not sure how much of a problem this
is, it seemed like the best way to stay DRY.

* Updates options for static_control method.

The `control_class` option is more appropriate for adding the static
class. This also will allow people to pass the `value` option to the
method call.

* Removes passing a block as an option from `static_control`.

Since the block will no longer behave the same as it did in previous
versions, it is probably best to remove it entirely. This way tests
will fail hard and make it easier to catch instead of silently behaving
differently.

This also will not overwrite a `control_class` if it is passed in as
an option by the user. There is a new test added to ensure this.

Updates to the CHANGELOG to reflect the breaking changes to this method.

* Removes empty extra list item from bugfixes.