Moves help text in horiz. forms under the control (#454)
* Updates help messages for horizontal forms test

Updates test to reflect the changes made in form_group method

* Moves help text in horiz. forms under the control 

When using horizontal forms the help text is placed inside the created row but as sibling of the cols for label and control. This codechange moves the help text inside the col of the control so it gets aligned with/under the control.