Document the gem release process; add chandler
Releasing to rubygems.org is as simple as `rake release`, but there are
some additional steps around it to make sure our documentation is up to
date, which I've documented in `RELEASING.md`.

I've also added the `chandler` gem (my own creation) that automatically
updates the GitHub Releases whenever we run `rake release`. This ensures
that people can find our release notes easily.

This is the process I used to recently publish 4.0.0.alpha1.