Allow overriding the default form_group class
I think the default wrapper class, "mb-3", is good, but there should be a way to omit it while still rendering the label and the rest of the form group content.

I realize this is an API change, but I think it is a good one which gives more power to the developers.  I also think it is in good alignment with the upstream bootstrap project.