Fix text alignment in alert message
When there is no error summary the alert message is wrapped in a p tag. This causes the message to be misaligned vertically. In the Bootstrap examples where there is single line of text, the text is not wrapped in a p tag: https://getbootstrap.com/docs/5.0/components/alerts/#examples