Remove the require expression for action text in lib/bootstrap_form.rb (#720)
This require was added in e6f5b09209147143ebdb83032a50e9758cfdc2ab to
add ActionText support.

Instead of eagerly loading the module, allow the Rails engine subsystem
to load the helpers as necessary.

Eagerly loading the module has the downside of defining the module
ActionText even when the engine is not in use. Defining the module like
this can "trick" other gems into erroneously loading their ActionText
support, which may lead to confusing and out of context error messages.